pyproject.toml
[tool.poetry]
name = "dobotapi"
version = "0.0.1"
description = "Python API for controlling Dobot Magician and it's addons"
authors = ["Stax124, JustABenda"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Stax124/DobotAPI"
repository = "https://github.com/Stax124/DobotAPI"
documentation = "stax124.github.io/DobotAPI/"

[tool.poetry.dependencies]
python = "^3.9"
prompt-toolkit = "^3.0.29"
pyserial = "^3.5"
coloredlogs = "^15.0.1"
"svg.path" = "^6.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
poethepoet = "^0.16.2"
black = {version = "^22.8.0", allow-prereleases = true}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
